Remove atlas/victory folders
Samsung decided to split up the source files for several key components into atlas and victory directories. This makes patching files difficult and also clutters the source tree and makes it significantly larger than it needs to be.

Victory is used for the Sprint Epic 4G SGS phone
Atlas is used for the Verizon Fascinate and other SCH-i500 devices (maybe other SGS phones as well)

The ultimate goal of this was to make patching and cherry-picking much simpler from a development standpoint. It has the side effect of slimming down the source tree and making things easier to work with as well.